requirements.txt
sanic==22.6.2
pyyaml
aiohttp
python-dotenv
